+++ demo/components/victory-bar-demo.js
@@ -114,8 +114,27 @@ export default class App extends React.Component {
     return (
       <div className="demo">
         <h1>VictoryBar</h1>
+        <ChartWrap>
+          <VictoryBar
+            horizontal
+            data={[
+              {x: 1, y: "Label 1"},
+              {x: 7, y: "Label 2"},
+              {x: 3, y: "Label 3"},
+              {x: 4, y: "Label 4"}
+            ]}
+          />
+        </ChartWrap>
+
+        <ChartWrap>
+          <VictoryBar horizontal
+            data={[ {x: 1, y: 20}, {x: 7, y: 40}, {x: 3, y: 60}, {x: 4, y: 80} ]}
+          />
+        </ChartWrap>
+
         <VictoryBar
           style={{
+            data: {width: 10},
             parent: parentStyle,
             labels: {angle: 45, verticalAnchor: "end", textAnchor: "end"}
           }}