+++ docs/victory-candlestick/docs.js
@@ -1,10 +1,11 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import Ecology from "ecology";
-import { merge, random } from "lodash";
 import Radium, { Style } from "radium";
+import { merge, random, range } from "lodash";
 import * as docgen from "react-docgen";
-import { VictoryLine, VictoryScatter } from "../../src/index";
+import { VictoryCandlestick, VictoryChart, VictoryAxis} from "../../src/index";
+import { VictoryLabel } from "victory-core";
 import { VictoryTheme, appendLinkIcon, ecologyPlaygroundLoading } from "formidable-landers";
 
 class Docs extends React.Component {
@@ -13,8 +14,13 @@ class Docs extends React.Component {
       <div>
         <Ecology
           overview={require("!!raw!./ecology.md")}
-          source={docgen.parse(require("!!raw!../../src/components/victory-line/victory-line"))}
-          scope={{ merge, random, React, ReactDOM, VictoryLine, VictoryScatter }}
+          source={docgen.parse(require(
+            "!!raw!../../src/components/victory-candlestick/victory-candlestick"
+            ))}
+          scope={{
+            merge, range, random, React, ReactDOM, VictoryLabel,
+            VictoryCandlestick, VictoryChart, VictoryAxis
+          }}
           playgroundtheme="elegant"
           customRenderers={merge(appendLinkIcon, ecologyPlaygroundLoading)}
         />