+++ src/components/victory-bar/bar.js
@@ -3,15 +3,22 @@ import React, { PropTypes } from "react";
 export default class Bar extends React.Component {
 
   static propTypes = {
-    index: PropTypes.number,
+    datum: PropTypes.object,
     events: PropTypes.object,
     horizontal: PropTypes.bool,
+    index: PropTypes.number,
+    role: PropTypes.string,
     scale: PropTypes.object,
     style: PropTypes.object,
-    datum: PropTypes.object,
     x: React.PropTypes.number,
     y: React.PropTypes.number,
-    y0: React.PropTypes.number
+    y0: React.PropTypes.number,
+    width: PropTypes.number,
+    padding: PropTypes.oneOfType([
+      PropTypes.number,
+      PropTypes.object
+    ]),
+    data: PropTypes.array
   };
 
   getVerticalBarPath(props, width) {
@@ -39,18 +46,21 @@ export default class Bar extends React.Component {
       this.getHorizontalBarPath(props, width) : this.getVerticalBarPath(props, width);
   }
 
+  renderBar(path, style, events) {
+    const { role } = this.props;
+    return (
+      <path d={path} style={style} role={role} shapeRendering="optimizeSpeed" {...events}/>
+    );
+  }
+
   render() {
     // TODO better bar width calculation
-    const barWidth = this.props.style && this.props.style.width || 8;
+    const { data, events, style, width} = this.props;
+    const padding = this.props.padding.left || this.props.padding;
+    const barWidth = style && style.width ||
+    0.3 * (width - 2 * padding) / data.length;
     const path = typeof this.props.x === "number" ?
       this.getBarPath(this.props, barWidth) : undefined;
-    return (
-      <path
-        {...this.props.events}
-        d={path}
-        style={this.props.style}
-        shapeRendering="optimizeSpeed"
-      />
-    );
+    return this.renderBar(path, style, events);
   }
 }