+++ src/components/victory-bar/victory-bar.js
@@ -16,7 +16,6 @@ const fallbackProps = {
   },
   style: {
     data: {
-      width: 8,
       padding: 6,
       stroke: "transparent",
       strokeWidth: 0,
@@ -322,7 +321,13 @@ export default class VictoryBar extends React.Component {
     * @example theme={VictoryTheme.grayscale}
     * http://www.github.com/FormidableLabs/victory-core/tree/master/src/victory-theme/grayscale.js
     */
-    theme: PropTypes.object
+    theme: PropTypes.object,
+    /**
+     * The groupComponent prop takes an entire component which will be used to
+     * create group elements for use within container elements. This prop defaults
+     * to a <g> tag on web, and a react-native-svg <G> tag on mobile
+     */
+    groupComponent: PropTypes.element
   };
 
   static defaultProps = {
@@ -334,7 +339,8 @@ export default class VictoryBar extends React.Component {
     standalone: true,
     x: "x",
     y: "y",
-    containerComponent: <VictoryContainer/>
+    containerComponent: <VictoryContainer/>,
+    groupComponent: <g/>
   };
 
   static getDomain = Domain.getDomainWithZero.bind(Domain);
@@ -366,13 +372,15 @@ export default class VictoryBar extends React.Component {
   }
 
   renderData(props) {
-    const { dataComponent, labelComponent } = props;
+    const { dataComponent, labelComponent, groupComponent } = props;
+    const { role } = VictoryBar;
     const barComponents = [];
     const barLabelComponents = [];
-    this.dataKeys.forEach((key) => {
+
+    this.dataKeys.forEach((key, index) => {
       const dataEvents = this.getEvents(props, "data", key);
       const dataProps = defaults(
-        {key: `bar-${key}`},
+        {index, key: `${role}-${key}`, role: `${role}-${index}`},
         this.getEventState(key, "data"),
         this.getSharedEventState(key, "data"),
         dataComponent.props,
@@ -384,7 +392,7 @@ export default class VictoryBar extends React.Component {
       )));
 
       const labelProps = defaults(
-        {key: `bar-label-${key}`},
+        {key: `${role}-label-${key}`},
         this.getEventState(key, "labels"),
         this.getSharedEventState(key, "labels"),
         labelComponent.props,
@@ -400,11 +408,8 @@ export default class VictoryBar extends React.Component {
     });
 
     if (barLabelComponents.length > 0) {
-      return (
-        <g>
-          {barComponents}
-          {barLabelComponents}
-        </g>
+      return React.cloneElement(
+        groupComponent, {}, ...barComponents, ...barLabelComponents
       );
     }
     return barComponents;
@@ -428,6 +433,14 @@ export default class VictoryBar extends React.Component {
     );
   }
 
+  renderGroup(children, style) {
+    return React.cloneElement(
+      this.props.groupComponent,
+      { role: "presentation", style},
+      children
+    );
+  }
+
   render() {
     const modifiedProps = Helpers.modifyProps(this.props, fallbackProps);
     const { animate, style, standalone } = modifiedProps;
@@ -438,7 +451,7 @@ export default class VictoryBar extends React.Component {
       ];
       return (
         <VictoryTransition animate={animate} animationWhitelist={whitelist}>
-          <VictoryBar {...modifiedProps}/>
+          {React.createElement(this.constructor, modifiedProps)}
         </VictoryTransition>
       );
     }
@@ -448,11 +461,7 @@ export default class VictoryBar extends React.Component {
 
     const baseStyles = Helpers.getStyles(style, styleObject, "auto", "100%");
 
-    const group = (
-      <g role="presentation" style={baseStyles.parent}>
-        {this.renderData(modifiedProps)}
-      </g>
-    );
+    const group = this.renderGroup(this.renderData(modifiedProps), baseStyles.parent);
 
     return standalone ? this.renderContainer(modifiedProps, group) : group;
   }