+++ src/components/victory-scatter/point.js
@@ -4,23 +4,24 @@ import pathHelpers from "./path-helpers";
 
 export default class Point extends React.Component {
   static propTypes = {
-    index: React.PropTypes.number,
     datum: PropTypes.object,
     events: PropTypes.object,
+    index: PropTypes.number,
+    role: PropTypes.string,
+    size: PropTypes.oneOfType([
+      PropTypes.number,
+      PropTypes.func
+    ]),
     symbol: PropTypes.oneOfType([
       PropTypes.oneOf([
         "circle", "diamond", "plus", "square", "star", "triangleDown", "triangleUp"
       ]),
       PropTypes.func
     ]),
-    size: PropTypes.oneOfType([
-      PropTypes.number,
-      PropTypes.func
-    ]),
     scale: PropTypes.object,
     style: PropTypes.object,
-    x: React.PropTypes.number,
-    y: React.PropTypes.number
+    x: PropTypes.number,
+    y: PropTypes.number
   };
 
   getPath(props) {
@@ -36,14 +37,14 @@ export default class Point extends React.Component {
     return pathFunctions[props.symbol].call(null, props.x, props.y, props.size);
   }
 
-  render() {
+  renderPoint(path, style, events) {
+    const { role } = this.props;
     return (
-      <path
-        {...this.props.events}
-        style={this.props.style}
-        d={this.getPath(this.props)}
-        shapeRendering="optimizeSpeed"
-      />
+      <path {...events} d={path} role={role} shapeRendering="optimizeSpeed" style={style}/>
     );
+  }
+
+  render() {
+    return this.renderPoint(this.getPath(this.props), this.props.style, this.props.events);
   }
 }