+++ src/components/victory-scatter/victory-scatter.js
@@ -342,7 +342,13 @@ export default class VictoryScatter extends React.Component {
     * @example theme={VictoryTheme.grayscale}
     * http://www.github.com/FormidableLabs/victory-core/tree/master/src/victory-theme/grayscale.js
     */
-    theme: PropTypes.object
+    theme: PropTypes.object,
+    /**
+     * The groupComponent prop takes an entire component which will be used to
+     * create group elements for use within container elements. This prop defaults
+     * to a <g> tag on web, and a react-native-svg <G> tag on mobile
+     */
+    groupComponent: PropTypes.element
   };
 
   static defaultProps = {
@@ -356,7 +362,8 @@ export default class VictoryScatter extends React.Component {
     y: "y",
     dataComponent: <Point/>,
     labelComponent: <VictoryLabel/>,
-    containerComponent: <VictoryContainer/>
+    containerComponent: <VictoryContainer/>,
+    groupComponent: <g/>
   };
 
   static getDomain = Domain.getDomain.bind(Domain);
@@ -389,13 +396,14 @@ export default class VictoryScatter extends React.Component {
   }
 
   renderData(props) {
-    const { dataComponent, labelComponent } = props;
+    const { dataComponent, labelComponent, groupComponent } = props;
+    const { role } = VictoryScatter;
     const pointComponents = [];
     const pointLabelComponents = [];
-    this.dataKeys.forEach((key) => {
+    this.dataKeys.forEach((key, index) => {
       const dataEvents = this.getEvents(props, "data", key);
       const dataProps = defaults(
-        {key: `scatter-${key}`},
+        {index, key: `${role}-${key}`, role: `${role}-${index}`},
         this.getEventState(key, "data"),
         this.getSharedEventState(key, "data"),
         dataComponent.props,
@@ -407,7 +415,7 @@ export default class VictoryScatter extends React.Component {
       )));
 
       const labelProps = defaults(
-        {key: `scatter-label-${key}`},
+        {key: `scatter-label-${key}`, index},
         this.getEventState(key, "labels"),
         this.getSharedEventState(key, "labels"),
         labelComponent.props,
@@ -422,11 +430,8 @@ export default class VictoryScatter extends React.Component {
     });
 
     if (pointLabelComponents.length > 0) {
-      return (
-        <g>
-          {pointComponents}
-          {pointLabelComponents}
-        </g>
+      return React.cloneElement(
+        groupComponent, {}, ...pointComponents, ...pointLabelComponents
       );
     }
     return pointComponents;
@@ -450,6 +455,14 @@ export default class VictoryScatter extends React.Component {
     );
   }
 
+  renderGroup(children, style) {
+    return React.cloneElement(
+      this.props.groupComponent,
+      { role: "presentation", style},
+      children
+    );
+  }
+
   render() {
     const modifiedProps = Helpers.modifyProps(this.props, fallbackProps);
     const { animate, style, standalone } = modifiedProps;
@@ -464,7 +477,7 @@ export default class VictoryScatter extends React.Component {
       ];
       return (
         <VictoryTransition animate={animate} animationWhitelist={whitelist}>
-          <VictoryScatter {...modifiedProps}/>
+          {React.createElement(this.constructor, modifiedProps)}
         </VictoryTransition>
       );
     }
@@ -475,11 +488,7 @@ export default class VictoryScatter extends React.Component {
 
     const baseStyles = Helpers.getStyles(style, styleObject, "auto", "100%");
 
-    const group = (
-      <g role="presentation" style={baseStyles.parent}>
-        {this.renderData(modifiedProps)}
-      </g>
-    );
+    const group = this.renderGroup(this.renderData(modifiedProps), baseStyles.parent);
 
     return standalone ? this.renderContainer(modifiedProps, group) : group;
   }